Initial Commit
This is my initial gist from Codepen which I used to create the HTML base for the FT Headlines website, drawing upon Origami components to give the website the look and feel of an FT product. The next stage is to integrate the Headline Access API to display the headlines and provide the search function.